navbar authorizations

+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\ShareController;
 
 /*
 |--------------------------------------------------------------------------
@@ -17,14 +18,17 @@ Route::get('/', function () {
     return view('welcome');
 });
 
-Route::middleware(['auth:sanctum', 'verified'])->get('/dashboard', function () {
-    return view('dashboard');
-})->name('dashboard');
+//Route::middleware(['auth:sanctum', 'verified'])->get('/dashboard', [ShareController::class,'welcome'])->name('dashboard');
+Route::get('/dashboard', [ShareController::class,'welcome'])->name('dashboard');
 
-Route::get('/shares', [ShareController::class,'index'])->middleware(['auth']);
+Route::get('/shares', [ShareController::class,'index']);
 Route::get('/share/show/{share:id}', [ShareController::class,'show']);
-Route::get('/share/create', [ShareController::class,'create']);
-Route::post('/share/create', [ShareController::class,'store']);
-Route::get('/share/edit/{share:id}',[ShareController::class,'edit']);
-Route::patch('/share/edit/{share:id}',[ShareController::class,'update']);
-Route::delete('/share/delete/{share:id}',[ShareController::class,'destroy']);
+Route::get('/share/create', [ShareController::class,'create'])->name('new')->middleware(['auth']);
+Route::post('/share/create', [ShareController::class,'store'])->middleware(['auth']);
+Route::get('/share/edit/{share:id}',[ShareController::class,'edit'])->middleware(['auth']);
+Route::patch('/share/edit/{share:id}',[ShareController::class,'update'])->middleware(['auth']);
+Route::delete('/share/delete/{share:id}',[ShareController::class,'destroy'])->middleware(['auth']);
+
+Route::get('/profile', [ShareController::class,'profile'])->name('profile');
+Route::get('/profile/{share:user_id}', [ShareController::class,'userprof'])->name('userprof');
+
